Move preset helpers to backend module
